feat(tests): add skipList to verifications
also makes verifications run on every `npm test`, not just if CI=1
