+++ package.json
@@ -8,8 +8,7 @@
     "new-template": "node _helpers/add-template.js",
     "add-dependency": "npm install $npm_config_package --no-package-lock --prefix $npm_config_template && npm install --save-dev $npm_config_package",
     "env-manifest": "node _helpers/generate-env-variable-manifest",
-    "ci:generate-manifest": "npm run env-manifest -- --write",
-    "ci:verify-templates": "CI=1 jest test/ci/verify-templates.test.js"
+    "ci:generate-manifest": "npm run env-manifest -- --write"
   },
   "author": "Twilio <open-source@twilio.com>",
   "license": "MIT",
